requirements.txt
biplist
pycryptodome